==> 03.SVM-RFE.R <==

library(e1071)   
library(kernlab)   
library(caret)      

set.seed(123)     

inputFile="diffGeneExp.txt"  

setwd(" ")

data=read.table(inputFile, header=T, sep="\t", check.names=F, row.names=1)
data=t(data)

group=gsub("(.*)\\_(.*)", "\\2", row.names(data))

Profile=rfe(x=data,
            y=as.numeric(as.factor(group)),  
            sizes = c(2,4,6,8, seq(10,40,by=3)),  
            rfeControl = rfeControl(functions = caretFuncs, method = "cv"),  
            methods="svmRadial") 

pdf(file="SVM-RFE.pdf", width=6, height=5.5)  
par(las=1)
x = Profile$results$Variables  
y = Profile$results$RMSE      
plot(x, y, xlab="Variables", ylab="RMSE (Cross-Validation)", col="darkgreen") 
lines(x, y, col="darkgreen")

wmin=which.min(y)  
wmin.x=x[wmin]
wmin.y=y[wmin]
points(wmin.x, wmin.y, col="blue", pch=16) 
text(wmin.x, wmin.y, paste0('N=',wmin.x), pos=2, col=2)  
dev.off() 

featureGenes=Profile$optVariables  
write.table(file="SVM-RFE.gene.txt", featureGenes, sep="\t", quote=F, row.names=F, col.names=F)  # 将特征写入txt文件
